requirements.txt
Django==3.0.3
mysqlclient==1.4.6
django-mysql==2.2.0
djangorestframework==3.11.0
djoser==2.0.3
djangorestframework-simplejwt==4.4.0
django-filter==2.2.0
django-cors-headers==3.2.1
drf-writable-nested==0.5.4
omise==0.8.1
coreapi==2.3.3